+++ src/Config/ConfigAwareInterface.php
@@ -1,9 +1,12 @@
 <?php
+
 /**
- * This file is part of the InaneTools package.
+ * Inane Tools
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
+ *
+ * PHP version 8.1
  *
  * @author Philip Michael Raab <philip@inane.co.za>
  * @package Inane\Config
@@ -11,7 +14,7 @@
  * @license MIT
  * @license https://inane.co.za/license/MIT
  *
- * @copyright 2015-2019 Philip Michael Raab <philip@inane.co.za>
+ * @copyright 2015-2022 Philip Michael Raab <philip@inane.co.za>
  */
 
 /**
@@ -64,18 +67,27 @@
  * 	}
  */
 
+declare(strict_types=1);
+
 namespace Inane\Config;
+
+use Laminas\Config\Config;
 
 /**
  * ConfigAwareInterface
  *
  * @package Inane\Config
- * @version 0.2.0
+ *
+ * @version 1.0.0
  */
 interface ConfigAwareInterface {
 
 	/**
-	 * @param Array $config
+	 * configuration
+	 *
+	 * @param array|\Inane\Config\Options|\Laminas\Config\Config $config configuration
+	 *
+	 * @return void
 	 */
-	public function setConfig($config);
+	public function setConfig(array|Options|Config $config): void;
 }